requirements.txt
psutil>=5.9.0
Jinja2>=3.0.0
plyer>=2.0.0
PyYAML>=6.0
pytest>=7.0.0
rich>=13.0.0
